Rev4/requirements.txt
langchain
langchain-openai
langchain-community
chromadb
pandas
numpy
scipy
colorama
python-dotenv
